+++ src/main/java/com/ddang/usedauction/point/domain/PointHistory.java
@@ -1,10 +1,23 @@
 package com.ddang.usedauction.point.domain;
 
 import com.ddang.usedauction.config.BaseTimeEntity;
-import com.ddang.usedauction.member.Member;
+import com.ddang.usedauction.member.domain.Member;
 import com.ddang.usedauction.point.type.PointType;
-import jakarta.persistence.*;
-import lombok.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -23,11 +36,11 @@ public class PointHistory extends BaseTimeEntity {
 
     // 충전 or 사용 포인트량
     @Column(nullable = false)
-    private int pointAmount;
+    private long pointAmount;
 
     // 현재 보유 포인트량
     @Column(nullable = false)
-    private int curPointAmount;
+    private long curPointAmount;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id", nullable = false)
